Добавляет решение для задания Модуляция
