+++ index.html
@@ -235,6 +235,7 @@
     </section>
   </template>
 
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
